+++ examples/boids/Cargo.toml
@@ -11,3 +11,4 @@ getrandom = { version = "0.2", features = ["js"] }
 rand = "0.8"
 serde = { version = "1.0", features = ["derive"] }
 yew = { path = "../../yew" }
+yew-services = { path = "../../yew-services" }
