+++ examples/crm/Cargo.toml
@@ -8,3 +8,4 @@ edition = "2018"
 serde = "1"
 serde_derive = "1"
 yew = { path = "../../yew" }
+yew-services = { path = "../../yew-services" }
