+++ examples/crm/src/main.rs
@@ -1,9 +1,9 @@
 use add_client::AddClientForm;
 use serde::{Deserialize, Serialize};
 use yew::format::Json;
-use yew::services::storage::Area;
-use yew::services::{DialogService, StorageService};
 use yew::{html, Component, ComponentLink, Html, ShouldRender};
+use yew_services::storage::Area;
+use yew_services::{DialogService, StorageService};
 
 mod add_client;
