+++ examples/game_of_life/Cargo.toml
@@ -14,3 +14,4 @@ log = "0.4"
 rand = "0.8"
 wasm-logger = "0.2"
 yew = { path = "../../yew" }
+yew-services = { path = "../../yew-services" }
