+++ examples/multi_thread/Cargo.toml
@@ -9,3 +9,4 @@ log = "0.4"
 serde = { version = "1.0", features = ["derive"] }
 wasm-logger = "0.2"
 yew = { path = "../../yew" }
+yew-services = { path = "../../yew-services" }
