+++ examples/todomvc/src/main.rs
@@ -1,10 +1,10 @@
 use state::{Entry, Filter, State};
 use strum::IntoEnumIterator;
 use yew::format::Json;
-use yew::services::storage::{Area, StorageService};
 use yew::web_sys::HtmlInputElement as InputElement;
 use yew::{classes, html, Component, ComponentLink, Html, InputData, NodeRef, ShouldRender};
 use yew::{events::KeyboardEvent, Classes};
+use yew_services::storage::{Area, StorageService};
 
 mod state;
