+++ examples/webgl/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 js-sys = "0.3"
 wasm-bindgen = "0.2"
 yew = { path = "../../yew" }
+yew-services = { path = "../../yew-services" }
 
 [dependencies.web-sys]
 version = "0.3"
