+++ yew-functional/Cargo.toml
@@ -14,6 +14,7 @@ description = "A framework for making client-side single-page apps"
 wasm-bindgen-test = "0.3.9"
 web-sys = "0.3.36"
 yew = { path = "../yew" }
+yew-services = { path = "../yew-services" }
 
 [dependencies]
 yew = { path = "../yew" }
