+++ yew-functional/tests/use_context_hook.rs
@@ -34,7 +34,7 @@ fn use_context_scoping_works() {
 
         fn run(_props: &Self::TProps) -> Html {
             if use_context::<ExampleContext>().is_some() {
-                yew::services::ConsoleService::log(&format!(
+                yew_services::ConsoleService::log(&format!(
                     "Context should be None here, but was {:?}!",
                     use_context::<ExampleContext>().unwrap()
                 ));
