+++ yew-router/Cargo.toml
@@ -31,7 +31,7 @@ web_sys = [
 ]
 
 [dependencies]
-yew = { version = "0.17.0", path = "../yew", features = ["services", "agent"], default-features= false, optional = true }
+yew = { version = "0.17.0", path = "../yew", features = ["agent"], default-features= false, optional = true }
 yew-router-macro = { version = "0.14.0", path = "../yew-router-macro" }
 yew-router-route-parser = { version = "0.14.0", path = "../yew-router-route-parser" }
