+++ yew-services/src/lib.rs
@@ -1,8 +1,12 @@
+#![recursion_limit = "512"]
+
 //! This module is a container of servies to interact with the external resources.
 //!
 //! It carries a similar role as subscriptions in Elm, but can be used directly
 //! from the `update` method.
 
+#[cfg(test)]
+pub(crate) mod callback_test_util;
 pub mod console;
 pub mod dialog;
 pub mod fetch;
