+++ yew-services/src/reader/web_sys.rs
@@ -1,8 +1,7 @@
 //! `web-sys` implementation for the reader service.
 
 use super::*;
-use crate::callback::Callback;
-use crate::services::Task;
+use crate::Task;
 #[doc(no_inline)]
 pub use ::web_sys::{Blob, File};
 use ::web_sys::{Event, FileReader};
@@ -10,6 +9,7 @@ use anyhow::{anyhow, Result};
 use gloo::events::EventListener;
 use js_sys::Uint8Array;
 use std::cmp;
+use yew::callback::Callback;
 
 impl ReaderService {
     /// Reads all bytes from a file and returns them with a callback.
