+++ yew-stdweb/examples/todomvc/src/lib.rs
@@ -6,7 +6,7 @@ use strum::IntoEnumIterator;
 use strum_macros::{EnumIter, ToString};
 use yew::events::IKeyboardEvent;
 use yew::format::Json;
-use yew::services::storage::{Area, StorageService};
+use yew_services::storage::{Area, StorageService};
 use yew::{html, Component, ComponentLink, Html, InputData, KeyPressEvent, ShouldRender};
 
 const KEY: &str = "yew.todomvc.self";
