+++ yew/Cargo.toml
@@ -122,7 +122,7 @@ rmp-serde = "0.15.0"
 bincode = "1"
 
 [features]
-default = ["services", "agent", "web_sys"]
+default = ["agent", "web_sys"]
 std_web = ["stdweb", "yew-macro/std_web"]
 web_sys = [
   "console_error_panic_hook",
@@ -136,7 +136,6 @@ doc_test = []
 wasm_test = []
 httpbin_test = []
 wasm_bench = []
-services = []
 agent = ["bincode"]
 yaml = ["serde_yaml"]
 msgpack = ["rmp-serde"]
